+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log (vs-deploy)
 
+## 5.34.0 (March 9th, 2017; use files of deployment for compilers)
+
+* added `useFilesOfDeployment` property for [compiler operations](https://github.com/mkloubert/vs-deploy/wiki/targetoperations#compile-), which indicates if files for deployment will be used as source, when not defined explicit
+
 ## 5.33.0 (March 4th, 2017; result button in status bar)
 
 * added more information for result button in status bar (s. [issue #42](https://github.com/mkloubert/vs-deploy/issues/42))